requirements.txt
fastapi
uvicorn
pydantic
pandas
python-dotenv
openpyxl
sqlalchemy
azure-communication-email
azure-communication-messages
